UPD: factorize MapieClassifier methods into several non-conformity scores + generic adaptation
