+++ mapie/conformity_scores/bounds/residuals.py
@@ -9,142 +9,11 @@ from sklearn.pipeline import Pipeline
 from sklearn.utils.validation import (check_is_fitted, check_random_state,
                                       indexable)
 
-from mapie._machine_precision import EPSILON
 from mapie._typing import ArrayLike, NDArray
-from mapie.conformity_scores import ConformityScore
+from mapie.conformity_scores import BaseRegressionScore
 
 
-class AbsoluteConformityScore(ConformityScore):
-    """
-    Absolute conformity score.
-
-    The signed conformity score = y - y_pred.
-    The conformity score is symmetrical.
-
-    This is appropriate when the confidence interval is symmetrical and
-    its range is approximatively the same over the range of predicted values.
-    """
-
-    def __init__(
-        self,
-        sym: bool = True,
-    ) -> None:
-        super().__init__(sym=sym, consistency_check=True)
-
-    def get_signed_conformity_scores(
-        self,
-        X: ArrayLike,
-        y: ArrayLike,
-        y_pred: ArrayLike,
-    ) -> NDArray:
-        """
-        Compute the signed conformity scores from the predicted values
-        and the observed ones, from the following formula:
-        signed conformity score = y - y_pred
-        """
-        return np.subtract(y, y_pred)
-
-    def get_estimation_distribution(
-        self,
-        X: ArrayLike,
-        y_pred: ArrayLike,
-        conformity_scores: ArrayLike
-    ) -> NDArray:
-        """
-        Compute samples of the estimation distribution from the predicted
-        values and the conformity scores, from the following formula:
-        signed conformity score = y - y_pred
-        <=> y = y_pred + signed conformity score
-
-        ``conformity_scores`` can be either the conformity scores or
-        the quantile of the conformity scores.
-        """
-        return np.add(y_pred, conformity_scores)
-
-
-class GammaConformityScore(ConformityScore):
-    """
-    Gamma conformity score.
-
-    The signed conformity score = (y - y_pred) / y_pred.
-    The conformity score is not symmetrical.
-
-    This is appropriate when the confidence interval is not symmetrical and
-    its range depends on the predicted values. Like the Gamma distribution,
-    its support is limited to strictly positive reals.
-    """
-
-    def __init__(
-        self,
-        sym: bool = False,
-    ) -> None:
-        super().__init__(sym=sym, consistency_check=False, eps=EPSILON)
-
-    def _check_observed_data(
-        self,
-        y: ArrayLike,
-    ) -> None:
-        if not self._all_strictly_positive(y):
-            raise ValueError(
-                f"At least one of the observed target is negative "
-                f"which is incompatible with {self.__class__.__name__}. "
-                "All values must be strictly positive, "
-                "in conformity with the Gamma distribution support."
-            )
-
-    def _check_predicted_data(
-        self,
-        y_pred: ArrayLike,
-    ) -> None:
-        if not self._all_strictly_positive(y_pred):
-            raise ValueError(
-                f"At least one of the predicted target is negative "
-                f"which is incompatible with {self.__class__.__name__}. "
-                "All values must be strictly positive, "
-                "in conformity with the Gamma distribution support."
-            )
-
-    @staticmethod
-    def _all_strictly_positive(
-        y: ArrayLike,
-    ) -> bool:
-        return not np.any(np.less_equal(y, 0))
-
-    def get_signed_conformity_scores(
-        self,
-        X: ArrayLike,
-        y: ArrayLike,
-        y_pred: ArrayLike,
-    ) -> NDArray:
-        """
-        Compute the signed conformity scores from the observed values
-        and the predicted ones, from the following formula:
-        signed conformity score = (y - y_pred) / y_pred
-        """
-        self._check_observed_data(y)
-        self._check_predicted_data(y_pred)
-        return np.divide(np.subtract(y, y_pred), y_pred)
-
-    def get_estimation_distribution(
-        self,
-        X: ArrayLike,
-        y_pred: ArrayLike,
-        conformity_scores: ArrayLike
-    ) -> NDArray:
-        """
-        Compute samples of the estimation distribution from the predicted
-        values and the conformity scores, from the following formula:
-        signed conformity score = (y - y_pred) / y_pred
-        <=> y = y_pred * (1 + signed conformity score)
-
-        ``conformity_scores`` can be either the conformity scores or
-        the quantile of the conformity scores.
-        """
-        self._check_predicted_data(y_pred)
-        return np.multiply(y_pred, np.add(1, conformity_scores))
-
-
-class ResidualNormalisedScore(ConformityScore):
+class ResidualNormalisedScore(BaseRegressionScore):
     """
     Residual Normalised score.
 
@@ -200,7 +69,8 @@ class ResidualNormalisedScore(ConformityScore):
         self.random_state = random_state
 
     def _check_estimator(
-        self, estimator: Optional[RegressorMixin] = None
+        self,
+        estimator: Optional[RegressorMixin] = None
     ) -> RegressorMixin:
         """
         Check if estimator is ``None``,
@@ -361,9 +231,10 @@ class ResidualNormalisedScore(ConformityScore):
 
     def get_signed_conformity_scores(
         self,
-        X: ArrayLike,
         y: ArrayLike,
-        y_pred: ArrayLike
+        y_pred: ArrayLike,
+        X: Optional[ArrayLike] = None,
+        **kwargs
     ) -> NDArray:
         """
         Computes the signed conformity score = (y - y_pred) / r_pred.
@@ -374,6 +245,8 @@ class ResidualNormalisedScore(ConformityScore):
         The learning is done with the log of the residual and later we
         use the exponential of the prediction to avoid negative values.
         """
+        assert not (X is None)  # TODO
+
         (X, y, y_pred,
          self.residual_estimator_,
          random_state) = self._check_parameters(X, y, y_pred)
@@ -418,9 +291,10 @@ class ResidualNormalisedScore(ConformityScore):
 
     def get_estimation_distribution(
         self,
-        X: ArrayLike,
         y_pred: ArrayLike,
-        conformity_scores: ArrayLike
+        conformity_scores: ArrayLike,
+        X: Optional[ArrayLike] = None,
+        **kwargs
     ) -> NDArray:
         """
         Compute samples of the estimation distribution from the predicted
@@ -433,6 +307,8 @@ class ResidualNormalisedScore(ConformityScore):
         ``conformity_scores`` can be either the conformity scores or
         the quantile of the conformity scores.
         """
+        assert not (X is None)  # TODO
+
         r_pred = self._predict_residual_estimator(X).reshape((-1, 1))
         if not self.prefit:
             return np.add(
