+++ mapie/regression/regression.py
@@ -12,13 +12,14 @@ from sklearn.utils import check_random_state
 from sklearn.utils.validation import _check_y, check_is_fitted, indexable
 
 from mapie._typing import ArrayLike, NDArray
-from mapie.conformity_scores import ConformityScore, ResidualNormalisedScore
+from mapie.conformity_scores import (BaseRegressionScore,
+                                     ResidualNormalisedScore)
+from mapie.conformity_scores.utils import check_regression_conformity_score
 from mapie.estimator.regressor import EnsembleRegressor
 from mapie.utils import (check_alpha, check_alpha_and_n_samples,
                          check_cv, check_estimator_fit_predict,
                          check_n_features_in, check_n_jobs, check_null_weight,
                          check_verbose, get_effective_calibration_samples)
-from mapie.conformity_scores.checks import check_conformity_score
 
 
 class MapieRegressor(BaseEstimator, RegressorMixin):
@@ -226,7 +227,7 @@ class MapieRegressor(BaseEstimator, RegressorMixin):
         n_jobs: Optional[int] = None,
         agg_function: Optional[str] = "mean",
         verbose: int = 0,
-        conformity_score: Optional[ConformityScore] = None,
+        conformity_score: Optional[BaseRegressionScore] = None,
         random_state: Optional[Union[int, np.random.RandomState]] = None,
     ) -> None:
         self.estimator = estimator
@@ -431,7 +432,7 @@ class MapieRegressor(BaseEstimator, RegressorMixin):
             self.method = "base"
         estimator = self._check_estimator(self.estimator)
         agg_function = self._check_agg_function(self.agg_function)
-        cs_estimator = check_conformity_score(
+        cs_estimator = check_regression_conformity_score(
             self.conformity_score, self.default_sym_
         )
         if isinstance(cs_estimator, ResidualNormalisedScore) and \
@@ -449,7 +450,7 @@ class MapieRegressor(BaseEstimator, RegressorMixin):
         # Casting
         cv = cast(BaseCrossValidator, cv)
         estimator = cast(RegressorMixin, estimator)
-        cs_estimator = cast(ConformityScore, cs_estimator)
+        cs_estimator = cast(BaseRegressionScore, cs_estimator)
         agg_function = cast(Optional[str], agg_function)
         X = cast(NDArray, X)
         y = cast(NDArray, y)
@@ -539,7 +540,7 @@ class MapieRegressor(BaseEstimator, RegressorMixin):
         # Compute the conformity scores (manage jk-ab case)
         self.conformity_scores_ = \
             self.conformity_score_function_.get_conformity_scores(
-                X, y, y_pred
+                y, y_pred, X=X
             )
 
         return self
@@ -639,16 +640,15 @@ class MapieRegressor(BaseEstimator, RegressorMixin):
                 check_alpha_and_n_samples(alpha_np, n)
 
             # Predict the target with confidence intervals
-            y_pred, y_pred_low, y_pred_up = \
-                self.conformity_score_function_.get_bounds(
-                    X,
-                    self.estimator_,
-                    self.conformity_scores_,
-                    alpha_np,
-                    ensemble=ensemble,
-                    method=self.method,
-                    optimize_beta=optimize_beta,
-                    allow_infinite_bounds=allow_infinite_bounds
-                )
+            outputs = self.conformity_score_function_.predict_set(
+                X, alpha_np,
+                estimator=self.estimator_,
+                conformity_scores=self.conformity_scores_,
+                ensemble=ensemble,
+                method=self.method,
+                optimize_beta=optimize_beta,
+                allow_infinite_bounds=allow_infinite_bounds
+            )
+            y_pred, y_pred_low, y_pred_up = outputs
 
             return np.array(y_pred), np.stack([y_pred_low, y_pred_up], axis=1)
